Add Filter for affiliates
Needing a fairly heavily customized leaderboard - and since we have no meta_query support (https://github.com/AffiliateWP/AffiliateWP/issues/2021) and no filters applied at the DB API level - a filter here would be super helpful!

A future PR (if it would be acceptable) might include an overrideable template for the output here. Currently, we'll have to rely on a bit of hackery on the affwp_show_leaderboard filter.

Cheers!